Compare end times in subscription filter enumerator
The filter was incorrectly checking the start time of data against the desired
end time for the algorithm. This led to the possibility of missing data points
at the end of the backtest.

One peculiarity here is that in UniverseSelection we round up the algorithm's
stated EndDate because QCAlgorithm subtracts a tick from midnight to prevent
rolling over to the next date. I'm not sure what other components depend on
that date and felt it best to leave it alone.