Remove copying of config/consolidators/price factors
Copying the price and sum of dividen fields in the copy constructor is very
incorrect since this constructor can be used to change the symbol, resolution,
data type and even the normalization mode. If we require this type of behavior,
then a separate (perhaps static) method should be introduced to perform this
type of extremely specific copy and also not include any of the override
parameters.