Remove daily universe selection for AddSecurity based universes
This forces the algorithm and data feed to synchronize once per day.

NOTE: It is somewhat expected to see changes in compounding annual return due
to different sampling as we no longer get samples at midnight each day. Also,
this change appears to showcase a bug in transaction synchronization with the
algo manager time loop and the ITransactionHandler.ProcessSynchronousEvents,
see LimitOrderRegressionAlgorithm